pyproject.toml
[project]
name = "opentelemetry-py"
version = "0.1.0"
description = "Playing with opentelemetry in python"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "opentelemetry-distro (>=0.54b1,<0.55)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false